requirements.txt
fastapi==0.115.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-genai==1.49.0
pydantic==2.9.2
pydantic-settings==2.12.0
pydantic_core==2.23.4
python-dotenv==1.2.1
uvicorn==0.30.6